requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# PDF processing
pdfplumber>=0.10.0
PyMuPDF>=1.23.0

# OCR (optional, for scanned PDFs)
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0

# Data processing
pandas>=2.1.0
openpyxl>=3.1.0

# Streamlit UI
streamlit>=1.28.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.6.0
